feat(nvim): migrate from lazy.nvim to native vim.pack
Migrated the entire Neovim configuration to use the native `vim.pack`
module (Neovim 0.12+). This change includes:

- Moving plugin configurations from `lua/modules/` to `plugin/` for
  automatic sourcing and reduced abstraction.
- Implementing a "configuration registry" pattern in `lua/core/pack.lua`
  to handle shared plugin configurations (e.g., `snacks.nvim`).
- Converting all plugin handles to full HTTPS URLs.
- Replacing `lazy.nvim`'s automatic lazy loading with eager loading and
  explicit keymap/autocmd definitions as requested.
- Managing dependency load order using numeric file prefixes in
  `plugin/` (e.g., `00-libs.lua`).
- Preserving all custom logic, including `mini.files` overrides and
  `overseer` templates.
- Adding `PackChanged` hooks for automatic `TSUpdate`.
- Removing `lazy-lock.json` and `lua/core/lazy-setup.lua`.
