+++ nvim/plugin/10-treesitter.lua
@@ -1,0 +1,22 @@
+vim.pack.add({
+    { src = "https://github.com/nvim-treesitter/nvim-treesitter", version = "main" },
+    "https://github.com/nvim-treesitter/nvim-treesitter-context",
+})
+
+-- Post-install hook for treesitter (simulating 'build = ":TSUpdate"')
+vim.api.nvim_create_autocmd('PackChanged', {
+  callback = function(ev)
+    local name, kind = ev.data.spec.name, ev.data.kind
+    if name == 'nvim-treesitter' and (kind == 'install' or kind == 'update') then
+      if not ev.data.active then vim.cmd.packadd('nvim-treesitter') end
+      vim.cmd('TSUpdate')
+    end
+  end
+})
+
+-- nvim-treesitter-context
+require("treesitter-context").setup({
+    max_lines = 1,
+    multiline_threshold = 1,
+    min_window_height = 20,
+})
